imgsearch/requirements.txt
configparser==3.5.0
Django==1.10.2
django-haystack==2.5.0
django-taggit==0.21.3
future==0.15.2
Pillow==3.4.2
requests==2.11.1
wheel==0.24.0
Whoosh==2.7.4
